refactor: move actions file
