+++ app/api/webhooks/clerk/route.ts
@@ -3,7 +3,7 @@ import {
   createUser,
   deleteUser,
   updateUser,
-} from "@/lib/database/models/actions/user.actions";
+} from "@/lib/database/actions/user.actions";
 import { clerkClient } from "@clerk/nextjs";
 import { WebhookEvent } from "@clerk/nextjs/server";
 import { headers } from "next/headers";
